Reworking internal files structure
