+++ inc/button.hpp
@@ -1,0 +1,12 @@
+#pragma once
+
+#include "frame.hpp"
+
+class Button : public Frame {
+public:
+	Button();
+
+	Button(const Button& other);
+
+	void show() const override;
+};
