+++ inc/cli.h
@@ -1,6 +1,8 @@
-#include <map>
-#include <string>
-#include <memory>
-#include "../inc/elements.hpp"
+#pragma once
 
-map<string, shared_ptr<Element>> open_elts_file(const string& file_name);
+// #include <map>
+// #include <string>
+// #include <memory>
+// #include "elements.hpp"
+
+// map<string, shared_ptr<Element>> open_elts_file(const string& file_name);
