+++ inc/functions.h
@@ -1,3 +1,5 @@
+#pragma once
+
 #include <cstdint>
 #include <string>
