+++ inc/window.hpp
@@ -1,0 +1,12 @@
+#pragma once
+
+#include "frame.hpp"
+
+class Window : public Frame {
+public:
+	Window();
+
+	// void init();
+	void set_size();
+	void show();
+};
