+++ src/button.cpp
@@ -1,0 +1,9 @@
+#include "../inc/button.hpp"
+
+Button::Button() : Frame() {}
+
+Button::Button(const Button& other) : Frame(other) {}
+
+void Button::show() const {
+	Frame::show();
+}
